+++ package.json
@@ -21,6 +21,7 @@
         "prettier": "^3.3.3"
     },
     "dependencies": {
+        "package.json": "^2.0.1",
         "primereact": "^10.8.4",
         "react-router-dom": "^6.27.0"
     }
